make uint64_t into size_t (same thing really) and remove a redundant cast

+++ src/Gpu/Device/Emulator/RuntimeManager/handleAlloc.cpp
@@ -8,7 +8,7 @@
 namespace Gpu::Device {
 
     Ipc::Response RuntimeManager::handleAlloc(const Ipc::Message& msg) {
-        const auto ptr = memory_.alloc(static_cast<std::size_t>(msg.size));
+        const auto ptr = memory_.alloc(msg.size);
 
         Ipc::Response res;
